initial working version

+++ index.js
@@ -1,19 +1,16 @@
 "use strict";
 
 const
-  Log = require('@dazn/lambda-powertools-logger');
-
-const
+  Log = require('@dazn/lambda-powertools-logger'),
   { provider } = require('./lib/provider'),
   { Logs } = require('./lib/cws'),
   config = require('./lib/config'),
-  { ok, extractUsers } = require('./lib/utils'),
-  LIMIT = 5;
+  { ok, extractUsers } = require('./lib/utils');
 
 
 const handler = async (event, ctx) => {
   Log.debug(`Event: ${JSON.stringify(event)}`);
-  let logs = new Logs(provider, LIMIT);
+  let logs = new Logs(provider, config.limit);
   let logGroups = await logs.fetchAllLogGroups();
   Log.info(`LogGroups: found ${logGroups.length}`);
   for (let group of logGroups) {
@@ -21,13 +18,14 @@ const handler = async (event, ctx) => {
     Log.debug(`LogGroup: analyze '${name}'`);
     let logStreams = await logs.fetchLogStreams(logGroupName);
     Log.debug(`LogStreams: found ${logStreams.length}`);
-    for (let r in logStreams) {
-      let { logStreamName } = logStreams[r]
-      await deleteLogStream(logGroupName, logStreamName);
+    if (config.delete) {
+      for (let r in logStreams) {
+        let { logStreamName } = logStreams[r];
+        await deleteLogStream(logGroupName, logStreamName);
+      }
     }
   }
-  Log.info(`fetched and analyzed`);
-  return ok(`OK`);
+  return ok('cleaned');
 };
 
 module.exports = {
